// server/src/models/types.ts
export type OrderState = 'Pendiente' | 'En preparación' | 'Enviado' | 'Entregado' | 'Cancelado';
export type DiscountType = 'porcentaje' | 'monto_fijo';
export type PaymentMethod = 'tarjeta' | 'yape' | 'paypal';
export type InvoiceType = 'boleta' | 'factura';
// ...
export interface CartItemInput {
  producto_id: string;
  variante_id?: string;
  cantidad: number;
}

export interface CartItemDetail {
  producto_id: string;
  variante_id: string;
  sku: string;
  nombre_producto: string;
  nombre_variante: string;
  color_nombre: string;
  color_hex: string;
  precio_unitario: number;
  precio_base: number;
  cantidad: number;
  subtotal: number;
  imagen: string;
  stock_disponible: number;
}

export interface CustomerData {
  nombre: string;
  email: string;
  telefono: string;
  tipo_documento: 'DNI' | 'CE' | 'RUC' | 'Pasaporte';
  numero_documento: string;
}

export interface ShippingAddress {
  departamento: string;
  provincia: string;
  distrito: string;
  direccion: string;
  referencia?: string;
  codigo_postal?: string;
}

export interface BillingData {
  tipo_comprobante: InvoiceType;
  ruc?: string;
  razon_social?: string;
  direccion_fiscal?: string;
}

export interface Order {
  id: string;
  codigo_orden: string;
  usuario_id?: string;
  datos_cliente: CustomerData;
  direccion_envio: ShippingAddress;
  metodo_envio: string;
  costo_envio: number;
  metodo_pago: PaymentMethod;
  estado_pago: 'completado' | 'simulado';
  subtotal: number;
  descuento: number;
  codigo_cupon?: string;
  total: number;
  tipo_comprobante: InvoiceType;
  datos_facturacion?: BillingData;
  estado_pedido: OrderState;
  items: CartItemDetail[];
  notas?: string;
  creado_en: string;
}